+++ src/contract/sent_transaction.ts
@@ -1,7 +1,11 @@
-import type { ContractClientOptions, Tx } from "./types";
-import { SorobanDataBuilder, SorobanRpc, TransactionBuilder } from "..";
+/* disable max-classes rule, because extending error shouldn't count! */
+/* eslint max-classes-per-file: 0 */
+import { SorobanDataBuilder, TransactionBuilder } from "@stellar/stellar-base";
+import type { ClientOptions, MethodOptions, Tx } from "./types";
+import { Server } from "../rpc/server"
+import { Api } from "../rpc/api"
 import { DEFAULT_TIMEOUT, withExponentialBackoff } from "./utils";
-import { AssembledTransaction } from "./assembled_transaction";
+import type { AssembledTransaction } from "./assembled_transaction";
 
 /**
  * A transaction that has been sent to the Soroban network. This happens in two steps:
@@ -18,33 +22,38 @@ import { AssembledTransaction } from "./assembled_transaction";
  *    `getTransactionResponse`.
  */
 export class SentTransaction<T> {
-  public server: SorobanRpc.Server;
+  public server: Server;
+
   public signed?: Tx;
+
   /**
    * The result of calling `sendTransaction` to broadcast the transaction to the
    * network.
    */
-  public sendTransactionResponse?: SorobanRpc.Api.SendTransactionResponse;
+  public sendTransactionResponse?: Api.SendTransactionResponse;
+
   /**
    * If `sendTransaction` completes successfully (which means it has `status: 'PENDING'`),
    * then `getTransaction` will be called in a loop for
    * {@link MethodOptions.timeoutInSeconds} seconds. This array contains all
    * the results of those calls.
    */
-  public getTransactionResponseAll?: SorobanRpc.Api.GetTransactionResponse[];
+  public getTransactionResponseAll?: Api.GetTransactionResponse[];
+
   /**
    * The most recent result of calling `getTransaction`, from the
    * `getTransactionResponseAll` array.
    */
-  public getTransactionResponse?: SorobanRpc.Api.GetTransactionResponse;
+  public getTransactionResponse?: Api.GetTransactionResponse;
 
   static Errors = {
-    SendFailed: class SendFailedError extends Error {},
-    SendResultOnly: class SendResultOnlyError extends Error {},
+    SendFailed: class SendFailedError extends Error { },
+    SendResultOnly: class SendResultOnlyError extends Error { },
+    TransactionStillPending: class TransactionStillPendingError extends Error { },
   };
 
   constructor(
-    public signTransaction: ContractClientOptions["signTransaction"],
+    public signTransaction: ClientOptions["signTransaction"],
     public assembled: AssembledTransaction<T>,
   ) {
     if (!signTransaction) {
@@ -52,7 +61,7 @@ export class SentTransaction<T> {
         "You must provide a `signTransaction` function to send a transaction",
       );
     }
-    this.server = new SorobanRpc.Server(this.assembled.options.rpcUrl, {
+    this.server = new Server(this.assembled.options.rpcUrl, {
       allowHttp: this.assembled.options.allowHttp ?? false,
     });
   }
@@ -62,20 +71,26 @@ export class SentTransaction<T> {
    * a `signTransaction` function. This will also send the transaction to the
    * network.
    */
-  static init = async <T>(
-    signTransaction: ContractClientOptions["signTransaction"],
-    assembled: AssembledTransaction<T>,
-  ): Promise<SentTransaction<T>> => {
+  static init = async <U>(
+    /** More info in {@link MethodOptions} */
+    signTransaction: ClientOptions["signTransaction"],
+    /** {@link AssembledTransaction} from which this SentTransaction was initialized */
+    assembled: AssembledTransaction<U>,
+  ): Promise<SentTransaction<U>> => {
     const tx = new SentTransaction(signTransaction, assembled);
-    return await tx.send();
+    const sent = await tx.send();
+    return sent;
   };
 
   private send = async (): Promise<this> => {
-    const timeoutInSeconds = this.assembled.options.timeoutInSeconds ?? DEFAULT_TIMEOUT
+    const timeoutInSeconds =
+      this.assembled.options.timeoutInSeconds ?? DEFAULT_TIMEOUT;
     this.assembled.built = TransactionBuilder.cloneFrom(this.assembled.built!, {
       fee: this.assembled.built!.fee,
       timebounds: undefined,
-      sorobanData: new SorobanDataBuilder(this.assembled.simulationData.transactionData.toXDR()).build()
+      sorobanData: new SorobanDataBuilder(
+        this.assembled.simulationData.transactionData.toXDR(),
+      ).build(),
     })
       .setTimeout(timeoutInSeconds)
       .build();
@@ -99,8 +114,11 @@ export class SentTransaction<T> {
 
     if (this.sendTransactionResponse.status !== "PENDING") {
       throw new SentTransaction.Errors.SendFailed(
-        "Sending the transaction to the network failed!\n" +
-          JSON.stringify(this.sendTransactionResponse, null, 2),
+        `Sending the transaction to the network failed!\n${JSON.stringify(
+          this.sendTransactionResponse,
+          null,
+          2,
+        )}`,
       );
     }
 
@@ -108,7 +126,7 @@ export class SentTransaction<T> {
 
     this.getTransactionResponseAll = await withExponentialBackoff(
       () => this.server.getTransaction(hash),
-      (resp) => resp.status === SorobanRpc.Api.GetTransactionStatus.NOT_FOUND,
+      (resp) => resp.status === Api.GetTransactionStatus.NOT_FOUND,
       timeoutInSeconds,
     );
 
@@ -116,21 +134,21 @@ export class SentTransaction<T> {
       this.getTransactionResponseAll[this.getTransactionResponseAll.length - 1];
     if (
       this.getTransactionResponse.status ===
-      SorobanRpc.Api.GetTransactionStatus.NOT_FOUND
+      Api.GetTransactionStatus.NOT_FOUND
     ) {
-      console.error(
+      throw new SentTransaction.Errors.TransactionStillPending(
         `Waited ${timeoutInSeconds} seconds for transaction to complete, but it did not. ` +
-          `Returning anyway. Check the transaction status manually. ` +
-          `Sent transaction: ${JSON.stringify(
-            this.sendTransactionResponse,
-            null,
-            2,
-          )}\n` +
-          `All attempts to get the result: ${JSON.stringify(
-            this.getTransactionResponseAll,
-            null,
-            2,
-          )}`,
+        `Returning anyway. Check the transaction status manually. ` +
+        `Sent transaction: ${JSON.stringify(
+          this.sendTransactionResponse,
+          null,
+          2,
+        )}\n` +
+        `All attempts to get the result: ${JSON.stringify(
+          this.getTransactionResponseAll,
+          null,
+          2,
+        )}`,
       );
     }
 
@@ -147,7 +165,8 @@ export class SentTransaction<T> {
         );
       }
 
-      // if "returnValue" not present, the transaction failed; return without parsing the result
+      // if "returnValue" not present, the transaction failed; return without
+      // parsing the result
       throw new Error("Transaction failed! Cannot parse result.");
     }